Crate debris_core

Modules

error

hir

llir

mir

objects

Structs

CompileContext

Config
Holds data about the user specified configuration

DatabaseStruct

DebrisObject
Objects are a central type for the compiler. Basically anything that can be assigned to a variable is an object This includes numbers, function, modules, and more It is possible to cast the ObjectPayload to its original value

InputsStorage
Representative struct for the query group.

ObjectRef
This struct is used to pass objects arround Objects references are read-only, so values that can be modified must be declared in a cell

Enums

BuildMode

Traits

Inputs

ObjectPayload
A trait for values that can be used as debris object payloads

Type Definitions

ObjectProperties