Module debris_core::class

This module defines classes. The concept of a class should be similar to the equivalent concept in other languages. Every value (object) has a class. Builtin values have a class with an associated type, while user defined classes, like structs, carry a reference to the concrete struct.

Structs

Class
A class combines ClassKind and corresponding methods. Due to the way classes get initialized, the methods have to be stored in a RefCell. For more info see crate::compile_context::TypeContext and debris_derive::object.

Enums

ClassKind

Type Definitions

ClassRef