Enum debris_core::class::ClassKind

pub enum ClassKind {
    Type(Type),
    Struct(StructRef),
    StructObject {
        strukt: StructRef,
        namespace: NamespaceIndex,
    },
    Tuple(TupleRef),
    TupleObject {
        tuple: TupleRef,
        namespace: NamespaceIndex,
    },
    Function {
        parameters: FunctionParameters,
        return_value: TypePattern,
    },
}

Variants

Type(Type)

Struct(StructRef)

StructObject

Fields of StructObject

strukt: StructRefnamespace: NamespaceIndex

Tuple(TupleRef)

TupleObject

Fields of TupleObject

tuple: TupleRefnamespace: NamespaceIndex

Function

Fields of Function

parameters: FunctionParametersreturn_value: TypePattern

Implementations

impl ClassKind

pub fn get_property(
&self,
arena: &NamespaceArena,
ident: &Ident
) -> Option<MirValue>

pub fn matches(&self, other: &ClassKind) -> bool

Returns whether the other class kind matches this class kind if this is interpreted as a pattern. For example, A struct object can match on a struct, if the underlying struct is equal.

pub fn matches_exact(&self, other: &ClassKind) -> bool

Returns whether other is the same class as self Behavior is the same as testing for equality, but StructObject does not compare the namespace, since it does not change the type

pub fn memory_ids<'a>(
&self,
buf: &mut Vec<ItemId>,
arena: &'a NamespaceArena,
default: ItemId
)

pub fn typ(&self) -> Type

pub fn is_never(&self) -> bool

Returns whether this class kind is of type Type::Never

pub fn runtime_encodable(&self) -> bool

Returns whether this type can be fully encoded at runtime.

pub fn pattern_runtime_encodable(&self) -> bool

Yeah…

pub fn comptime_encodable(&self) -> bool

Returns whether this type can be fully encoded at compile time.

pub fn matches_type(&self, typ: Type) -> bool

pub fn is_bool(&self) -> bool

pub fn is_function(&self) -> bool

Returns true if the class_kind is ClassKind::Function.