Struct debris_core::class::Class

pub struct Class {
    pub kind: ClassKind,
    pub properties: RefCell<ObjectProperties>,
}

A class combines ClassKind and corresponding methods. Due to the way classes get initialized, the methods have to be stored in a RefCell. For more info see crate::compile_context::TypeContext and debris_derive::object.

Fields
kind: ClassKindproperties: RefCell<ObjectProperties>
Implementations

impl Class

pub fn new_empty(kind: ClassKind) -> Self

pub fn matches(&self, other: &Class) -> bool

pub fn matches_exact(&self, other: &Class) -> bool

Returns whether the other class has the same type as this class.

pub fn get_property(
&self,
arena: &NamespaceArena,
ident: &Ident
) -> Option<MirValue>

pub fn set_property(&self, ident: Ident, obj_ref: ObjectRef)