Trait debris_core::function_interface::ValidReturnType

pub trait ValidReturnType {
    fn into_result(
        self, 
        ctx: &mut FunctionContext<'_, '_, '_, '_>
    ) -> LangResult<ObjectRef>;

    fn get_class(ctx: &CompileContext) -> Option<Rc<Class>>;
}

Trait used for converting any valid return value into a LangResult

Required methods

fn into_result(
self,
ctx: &mut FunctionContext<'_, '_, '_, '_>
) -> LangResult<ObjectRef>

fn get_class(ctx: &CompileContext) -> Option<Rc<Class>>

If possible, returns the type of the return value