Struct debris_core::llir::llir_builder::LlirBuilder

pub struct LlirBuilder<'llir, 'ctx, 'arena> {
    pub context: LlirContext<'ctx>,
    pub(super) current_function: BlockId,
    pub arena: &'arena mut NamespaceArena,
    pub nodes: PeepholeOptimizer,
    pub mir_contexts: &'ctx MirContextMap<'ctx>,
    pub llir_helper: &'llir mut LlirFunctions,
}

Fields
context: LlirContext<'ctx>current_function: BlockIdarena: &'arena mut NamespaceArenanodes: PeepholeOptimizermir_contexts: &'ctx MirContextMap<'ctx>llir_helper: &'llir mut LlirFunctions
Implementations

impl<'ctx, 'arena, 'llir> LlirBuilder<'llir, 'ctx, 'arena>

pub fn new(
context: &'ctx MirContext<'ctx>,
arena: &'arena mut NamespaceArena,
mir_contexts: &'ctx MirContextMap<'ctx>,
llir_helper: &'llir mut LlirFunctions
) -> Self

Creates a new LLIRBuilder. when building, populates the llir_helper struct.

pub fn build(self) -> Result<ObjectRef>

Populates the llir_functions vector

Returns the last expression and the id of the function

pub fn context_id(&self) -> ContextId

fn emit(&mut self, node: Node)

pub fn get_object(&self, value: &MirValue) -> ObjectRef

Converts a MirValue into an ObjectRef

If the value is not computed, return the default value

pub fn get_object_or_none(&self, value: &MirValue) -> Option<ObjectRef>

pub fn get_object_by_id(&self, id: ItemId) -> Option<ObjectRef>

Returns an object that belongs to this id. None if the object is still a template

pub fn set_object(&mut self, value: ObjectRef, id: ItemId)

Updates the template with this id to an object

fn replace_object(&mut self, value: ObjectRef, id: ItemId)

pub fn visit_context(&mut self, id: ContextId) -> Result<(BlockId, ObjectRef)>

Visits the context and optionally generates it. Returns the id and the return value