Enum debris_core::llir::llir_nodes::Node

pub enum Node {
    FastStore(FastStore),
    FastStoreFromResult(FastStoreFromResult),
    BinaryOperation(BinaryOperation),
    Call(Call),
    Condition(Condition),
    Branch(Branch),
    Execute(ExecuteRaw),
    Write(WriteMessage),
    Nop,
}

Any node

Variants

FastStore(FastStore)

FastStoreFromResult(FastStoreFromResult)

BinaryOperation(BinaryOperation)

Call(Call)

Condition(Condition)

Branch(Branch)

Execute(ExecuteRaw)

Write(WriteMessage)

Nop

Does nothing

Implementations

impl Node

pub fn iter<F>(&self, func: &mut F) where
F: FnMut(&Node),

Iterates over this node and all other nodes that this node contains. Changing this function also requires changing Node::iter_mut

pub fn iter_mut<F>(&mut self, func: &mut F) where
F: FnMut(&mut Node),

A copy of the above function, but with mutability enabled …

pub fn is_effect_free<'a, 'b: 'a, T>(&'a self, function_map: &'b T) -> bool where
T: Index<&'a BlockId, Output = Function>,

Returns whether this command has no side effect. sometimes its easier to just to restrict the parameter to a specifc type…

pub fn variable_accesses<F: FnMut(VariableAccess<'_>)>(&self, visitor: &mut F)

Accepts a callback function as argument and calls it with every variable accessed by this node.

pub fn variable_accesses_mut<F: FnMut(VariableAccessMut<'_>)>(
&mut self,
visitor: &mut F
)

See variable_accesses.

pub fn writes_to(&self, item_id: &ItemId) -> bool

Whether this node could modify item_id.

pub fn reads_from(&self, item_id: &ItemId) -> bool

Whether this node has a read-dependency on item_id