Module debris_core::mir

Mid-level intermediate representation

Once the high lever representation has be created, it can be lowered into a mir. The mir is quite simplistic and focuses on control flow and types.

An input of let a = 3 * 5 + 2 would be represented as a bunch of FunctionCall(...) and look somewhat like this:

Call(
    StaticInt.*,
    parameters: [StaticInt(3), StaticInt(5)],
    return_val: {id: 1, type: StaticInt}
);

Call(
    StaticInt.+,
    parameters: [{id: 1, type: StaticInt}, StaticInt(2)],
    return_val {id: 2, type: StaticInt}
);

The actual value of the returned static ints will be lazily evaluated at llir

A context is used to keep track of all variables that have ben created, as well as their identifiers.

Modules

mir_builder

mir_context

mir_context_kind

mir_control_flow

mir_impl

mir_nodes

mir_visitor

Structs

CachedFunctionSignature

ContextId

Mir

A Mid-level intermediate representation

MirBranchIf

Acts like MirGotoContext, if the condition is equal to true

MirBuilder

Visits the hir and creates a mir from it

MirCall

A function call to api functions

MirContext

Keeps track of single context, which can be a function, a loops or something else. As a rule of thumb, everything which has its own namespace is a context.

MirContextInfo

Helper struct which can hold mutable references to the arena and the context

MirContextMap

MirGotoContext

Calls a specific block of a context

MirJumpLocation

Command to start a new Block withing this context This allows jumps to this specific position

MirReturnValue

Sets the return value for a given context

MirUpdateValue

Overrides the old value in the namespace struct with new_value.

NamespaceArena

NamespaceIndex

ReturnValues

Enums

ContextKind

Contains all possible ‘kinds’ of contexts. This is for example used to determine wt which context to return when evaluating a return statement. Another use case is finding out, whether this context can be evaluated at comptime.

ControlFlowMode

The mode how a MirContext exits

MirNode

Any node that can be part of the mir representation

MirValue

Any value that is used in the mir compilation and also in the llir

Traits

MirVisitor