Enum debris_core::mir::MirNode

pub enum MirNode {
    Call(MirCall),
    GotoContext(MirGotoContext),
    JumpLocation(MirJumpLocation),
    UpdateValue(MirUpdateValue),
    ReturnValue(MirReturnValue),
    BranchIf(MirBranchIf),
}

Any node that can be part of the mir representation

Variants

Call(MirCall)

A function call which goes to a new block and initializes the parameters. How function calls work will probably change in the future.

GotoContext(MirGotoContext)

Goes to a specific jump location.

JumpLocation(MirJumpLocation)

Marks a jump location. A jump location can be seen as a ‘checkpoint’.

UpdateValue(MirUpdateValue)

Updates the value of a variable. This is the only node which may write to variables.

ReturnValue(MirReturnValue)

Sets the return value for this context. Note that this does not have to belong to a return-statement.

BranchIf(MirBranchIf)

Branches if a condition evaluates to true