Struct debris_core::mir::mir_context::MirContextInfo

pub struct MirContextInfo<'a, 'code> {
    pub context: &'a mut MirContext<'code>,
    pub arena: &'a mut NamespaceArena,
}

Helper struct which can hold mutable references to the arena and the context

Fields
context: &'a mut MirContext<'code>arena: &'a mut NamespaceArena
Implementations

impl<'a> MirContextInfo<'a, '_>

pub fn push(self, node: MirNode)

Pushes a node to the current context

pub fn add_unique_value(
self,
ident: Ident,
value: MirValue,
span: Span
) -> Result<()>

Adds a value to the current namespace. Fails if the ident is already associated with a value

pub fn add_value(self, ident: Ident, value: MirValue, span: Span)

Adds a value to the current value. Does not complain if there already exists a value with the same name (For example used for function parameters)

pub fn add_anonymous_template(self, class: Rc<Class>) -> MirValue

Adds the class as an anonymous template and returns it as a MirValue

pub fn declare_as_variable(self, id: ItemId, span: Span)

pub fn convert_into_template(self, item: ItemId)

pub fn get_from_spanned_ident(
self,
spanned_ident: &SpannedIdentifier
) -> Result<&'a MirValue>

Returns the value that is associated with this ident

pub fn resolve_path(self, path: &IdentifierPath) -> Result<AccessedProperty>

Follows this path

pub fn register_function_call(
self,
function: ObjectRef,
parameters: Vec<MirValue>,
parent: Option<MirValue>,
span: Span
) -> Result<(MirValue, MirNode)>

Adds a function call to the current context