Struct debris_core::mir::MirBuilder

pub struct MirBuilder<'a, 'ctx> {
    mir: &'a mut Mir<'ctx>,
    compile_context: &'ctx CompileContext,
    context_stack: ContextStack,
    hir_modules: &'a [HirModule],
    visited_functions: FxHashMap<Span, Rc<CachedFunctionSignature>>,
    hir_function_blocks: Vec<&'a HirBlock>,
    pub main_context: Option<ContextId>,
    pub ticking_contexts: Vec<ContextId>,
    global_context: ContextId,
}

Visits the hir and creates a mir from it

Fields
mir: &'a mut Mir<'ctx>compile_context: &'ctx CompileContextcontext_stack: ContextStackhir_modules: &'a [HirModule]

A list of all modules that are imported

visited_functions: FxHashMap<Span, Rc<CachedFunctionSignature>>

Cache for all function definitions that were already visited

hir_function_blocks: Vec<&'a HirBlock>main_context: Option<ContextId>

The first visited function gets marked as the main function

ticking_contexts: Vec<ContextId>

Contexts which got scheduled to run every tick

global_context: ContextId

The global, empty context which only contains modules

Implementations

impl<'a> MirBuilder<'a, '_>

pub fn visit(&mut self, block: &'a HirBlock) -> Result<MirValue>

Entry point for visiting the hir

fn visit_object(&mut self, object: &'a HirObject) -> Result<MirValue>

fn visit_struct(&mut self, hir_struct: &'a HirStruct) -> Result<MirValue>

fn visit_module(&mut self, module: &'a HirModule) -> Result<MirValue>

fn visit_block(&mut self, block: &'a HirBlock) -> Result<MirValue>

fn visit_import(&mut self, import: &'a HirImport) -> Result<MirValue>

fn visit_control_flow(
&mut self,
control_flow: &'a HirControlFlow
) -> Result<MirValue>

fn visit_function(&mut self, function: &'a HirFunction) -> Result<MirValue>

fn visit_conditional_branch(
&mut self,
branch: &'a HirConditionalBranch
) -> Result<MirValue>

fn visit_struct_initialization(
&mut self,
struct_instantiation: &'a HirStructInitialization
) -> Result<MirValue>

fn visit_tuple_initialization(
&mut self,
tuple_initialization: &'a HirTupleInitialization
) -> Result<MirValue>

fn visit_infinite_loop(
&mut self,
infinite_loop: &'a HirInfiniteLoop
) -> Result<MirValue>

fn visit_statement(&mut self, statement: &'a HirStatement) -> Result<MirValue>

fn visit_expression(
&mut self,
expression: &'a HirExpression
) -> Result<MirValue>

fn visit_function_call(
&mut self,
function_call: &'a HirFunctionCall
) -> Result<MirValue>

fn visit_variable_declaration(
&mut self,
variable_declaration: &'a HirVariableInitialization
) -> Result<MirValue>

fn visit_variable_update(
&mut self,
variable_update: &'a HirVariableUpdate
) -> Result<MirValue>

fn visit_const_value(
&mut self,
const_value: &'a HirConstValue
) -> Result<MirValue>

impl<'a, 'ctx> MirBuilder<'a, 'ctx>

pub fn new(
mir: &'a mut Mir<'ctx>,
hir_modules: &'a [HirModule],
modules: &[ModuleFactory],
compile_context: &'ctx CompileContext,
code: CodeRef<'ctx>
) -> Self

fn current_context(&self) -> ContextId

fn visit_block_local(&mut self, block: &'a HirBlock) -> Result<MirValue>

Visits a block without creating a new context

fn push_existing_context(&mut self, id: ContextId)

fn add_context(&mut self, span: Span, kind: ContextKind) -> ContextId

Creates a new context and pushes it to the top

fn add_context_after(
&mut self,
ancestor: ContextId,
span: Span,
kind: ContextKind
) -> ContextId

Creates a new context that is not successor of the current, but successor of a specific context

fn pop_context(&mut self) -> &mut MirContext<'ctx>

fn get_jump_target(
&self,
context_id: ContextId,
default: (ContextId, usize)
) -> Option<(ContextId, usize)>

Calculates the next jump target for context_id

fn push(&mut self, mir_node: MirNode)

Adds a mir node to the current context

fn add_mutable_value(&mut self, value: ObjectRef) -> MirValue

Adds a mutable concrete value to the current namespace. Returns the template

fn promote_runtime(&mut self, value: MirValue, span: Span) -> Result<MirValue>

Promotes a comptime value to its runtime variant

fn call_context(&mut self, context_id: ContextId, span: Span)

Generates a function call to that context. After the other context was executed, the current context will continue to run normally.

fn runtime_context(
&self,
stopping_context_id: ContextId
) -> Option<(ContextId, Span)>

Returns whether a variable declared in stopping_context is considered comptime. A variable is considered comptime, if every context in the context_stack from where the variable is declared to the current context is comptime.

fn call_function(
&mut self,
object: ObjectRef,
parent: Option<MirValue>,
parameters: Vec<MirValue>,
span: Span
) -> Result<MirValue>

Calls a function in the current context

fn instantiate_native_function(
&mut self,
parent: Option<MirValue>,
function_sig: &ObjNativeFunctionSignature,
parameters: &mut Vec<MirValue>,
span: Span
) -> Result<(ObjectRef, MirValue)>

Instantiates a native function signature and returns a NativeFunction object and its exact return type

fn try_clone_if_variable(
&mut self,
value: MirValue,
span: Span
) -> Result<MirValue>

Tries to clone this value or returns the original value if it does not need to be cloned

fn try_clone(
&mut self,
class: Rc<Class>,
id: ItemId,
span: Span
) -> Result<MirValue>

Tries to clone the value or returns it unmodified if that value cannot be cloned

fn update_return_value(
&mut self,
context_id: ContextId,
value: MirValue,
span: Span
) -> Result<usize>

fn get_type_pattern(&self, path: &HirTypePattern) -> Result<TypePattern>

fn assign_to_pattern(
&mut self,
pattern: &HirVariablePattern,
value: MirValue,
value_span: Span,
assignment_mode: PatternAssignmentMode
) -> Result<()>

Assigns value to pattern. If the pattern is a tuple pattern, then value is required to be a tuple.

fn accept_function_parameters(
&mut self,
expected: &[FunctionParameterDefinition],
parameters: &mut Vec<MirValue>,
parent: Option<MirValue>,
span: Span
) -> bool

Matches the parameters against the expected parameters. This operation can modify the given parameter vector.

fn register_ticking_function(
&mut self,
parameters: Vec<MirValue>,
span: Span
) -> Result<MirValue>

Compiler implemented debris function. Adds a function to the list of ticking functions.

impl<'code> MirBuilder<'_, 'code>

Implements functionality for working with the attributes mutably

pub fn context_mut(&mut self) -> &mut MirContext<'code>

Returns a mutable reference to the current context

pub fn context(&self) -> &MirContext<'_>

Returns a shared reference to the current context

pub fn context_info(&mut self) -> MirContextInfo<'_, 'code>

Returns a helper struct that contains both a context and the arena

pub fn namespace_mut(&mut self) -> &mut Namespace

Returns a mutable reference to the current namespace

pub fn arena_mut(&mut self) -> &mut NamespaceArena

Returns a mutable reference to the global arena

pub fn arena(&self) -> &NamespaceArena

Returns a shared reference to the global arena