Struct debris_core::mir::MirContext

pub struct MirContext<'ctx> {
    pub compile_context: &'ctx CompileContext,
    pub span: Span,
    pub kind: ContextKind,
    pub id: ContextId,
    pub nodes: Vec<MirNode>,
    pub control_flow: ControlFlowMode,
    pub return_values: ReturnValues,
    // some fields omitted
}

Keeps track of single context, which can be a function, a loops or something else. As a rule of thumb, everything which has its own namespace is a context.

Fields
compile_context: &'ctx CompileContext

A ref to the global compile context

span: Spankind: ContextKindid: ContextId

The context id and the id of the corresponding namespace

nodes: Vec<MirNode>

All mir nodes that are emitted

control_flow: ControlFlowMode

Which control flow mode is used for this context Most context just do nothing on end, but it is also possible to return from a function or to break from a loop

return_values: ReturnValues

The values returned by this function

Implementations

impl<'ctx> MirContext<'ctx>

pub fn new(
arena: &mut NamespaceArena,
ancestor_context: Option<ContextId>,
compile_context: &'ctx CompileContext,
span: Span,
kind: ContextKind
) -> Self

Creates a new context

If the ancestor index is not None, then this context will be a child of the ancestor and have access to its namespace

pub fn namespace_mut<'a>(
&self,
arena: &'a mut NamespaceArena
) -> &'a mut Namespace

pub fn namespace<'a>(&self, arena: &'a NamespaceArena) -> &'a Namespace

pub fn next_jump_location(&mut self) -> usize

Increments the jump location counter and returns it

pub fn register(&mut self, arena: &mut NamespaceArena, module: ObjModule)

Loads the module into this scope

pub fn register_members(
&mut self,
arena: &mut NamespaceArena,
module: ObjModule
)

Loads every member of the module (but not the module itself) into this scope

pub fn register_function_call(
&mut self,
arena: &mut NamespaceArena,
function: ObjectRef,
parameters: Vec<MirValue>,
parent: Option<MirValue>,
span: Span
) -> Result<(MirValue, MirNode)>

Creates a function call

pub fn add_value(
&self,
arena: &mut NamespaceArena,
ident: Ident,
value: MirValue,
span: Span
) -> Result<()>

Adds a value that corresponds to ident

If the value is a template, it is already added and only a namespace entry has to be created. Returns Err if the ident is already defined.

pub fn add_anonymous_object(
&self,
arena: &mut NamespaceArena,
object: ObjectRef
) -> MirValue

Adds an anonymous object and returns a ref MirValue

pub fn add_anonymous_template(
&self,
arena: &mut NamespaceArena,
class: GenericClassRef
) -> MirValue

Adds an anonymous value that is usually used temporarily

Returns the template as a MirValue.

pub fn get_from_spanned_ident<'a>(
&self,
arena: &'a NamespaceArena,
spanned_ident: &SpannedIdentifier
) -> Result<&'a MirValue>

Looks up the value that corresponds to this ident

pub fn get_ident(&self, spanned_ident: &SpannedIdentifier) -> Ident

Returns an ident from a span

pub fn get_type_pattern(
&self,
spanned_ident: &SpannedIdentifier
) -> Result<TypePattern>

Returns the type pattern that belongs to a single spanned identifier

pub fn resolve_path(
&self,
arena: &NamespaceArena,
path: &IdentifierPath
) -> Result<AccessedProperty>

Resolves the accessor and returns the accessed element