Module debris_core::objects

Structs

CallbackFunction
Wrapper, so traits like Eq can be implemented

DynInt

FunctionContext
The context which gets passed to a function

FunctionSignature
A signature describing a single overload of a function

FunctionSignatureMap
A collection of overloads for the same function As of right now, I only want to allow overloads for special functions (binary operations), but this can support more general functions as well

ModuleFactory

ObjectFunction
A function object Has a map of available signatures The call parameters are unique identifiers for every signature

ObjectModule
A module object. contains other modules

ObjectString

ObjectType

StaticInt

Type Definitions

TypeRef