Trait debris_core::Inputs

pub trait Inputs: Database + HasQueryGroup<InputsStorage> + Database {
    fn input_file(&self, key0: InputFile) -> String;

    fn set_input_file(&mut self, key0: InputFile, value__: String);

    fn set_input_file_with_durability(
        &mut self, 
        key0: InputFile, 
        value__: String, 
        durability__: Durability
    );

    fn extern_modules(&self) -> Rc<[ModuleFactory]>;

    fn set_extern_modules(&mut self, value__: Rc<[ModuleFactory]>);

    fn set_extern_modules_with_durability(
        &mut self, 
        value__: Rc<[ModuleFactory]>, 
        durability__: Durability
    );

    fn compile_context(&self) -> Rc<CompileContext>;

    fn set_compile_context(&mut self, value__: Rc<CompileContext>);

    fn set_compile_context_with_durability(
        &mut self, 
        value__: Rc<CompileContext>, 
        durability__: Durability
    );

    fn input_text(&self, key0: InputFile) -> CodeRef;

    fn config(&self, key0: InputFile) -> Rc<Config>;
}

Required methods

fn input_file(&self, key0: InputFile) -> String

fn set_input_file(&mut self, key0: InputFile, value__: String)

Set the value of the input_file input.

See input_file for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_input_file_with_durability(
&mut self,
key0: InputFile,
value__: String,
durability__: Durability
)

Set the value of the input_file input and promise that its value will never change again.

See input_file for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn extern_modules(&self) -> Rc<[ModuleFactory]>

fn set_extern_modules(&mut self, value__: Rc<[ModuleFactory]>)

Set the value of the extern_modules input.

See extern_modules for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_extern_modules_with_durability(
&mut self,
value__: Rc<[ModuleFactory]>,
durability__: Durability
)

Set the value of the extern_modules input and promise that its value will never change again.

See extern_modules for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn compile_context(&self) -> Rc<CompileContext>

fn set_compile_context(&mut self, value__: Rc<CompileContext>)

Set the value of the compile_context input.

See compile_context for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_compile_context_with_durability(
&mut self,
value__: Rc<CompileContext>,
durability__: Durability
)

Set the value of the compile_context input and promise that its value will never change again.

See compile_context for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn input_text(&self, key0: InputFile) -> CodeRef

fn config(&self, key0: InputFile) -> Rc<Config>

Loading content...