Trait debris_core::ObjectPayload

pub trait ObjectPayload: ValidPayload {
    fn memory_layout(&self) -> &MemoryLayout;

    fn create_class(&self, ctx: &CompileContext) -> Rc<Class> { ... }

    fn get_property(&self, _: &NamespaceArena, _: &Ident) -> Option<MirValue> { ... }

    fn as_function(&self) -> Option<&ObjFunction> { ... }
}

A trait for values that can be used as debris object payload

The private AsAny trait is auto-implemented

Required methods

fn memory_layout(&self) -> &MemoryLayout

Returns the memory layout of this specific object This method is usually only called once

Provided methods

fn create_class(&self, ctx: &CompileContext) -> Rc<Class>

The class specific to this object. Contains additionally to the class generics and the memory layout

Per default the default class of the object type

fn get_property(&self, _: &NamespaceArena, _: &Ident) -> Option<MirValue>

May be overwritten by distinct payloads which carry properties

fn as_function(&self) -> Option<&ObjFunction>

Interface any object can implement to be treated as a function