Crate debris_core

Core implementation of the debris compiler

This crate contains most of the compilation logic.

Compiling a debris script uses three different intermediate representations: First of all, the text is parsed and a ast-like structure gets produced. This happens in the hir module.

Subsequently, this structure gets lowered into a mid-level intermediate representation. The mir contains debris_objects.

At last the mir will be converted into a low-level intermediate representation. The llir is quite similar to actual minecraft command syntax and contains various nodes for scoreboard operations and more.

A very important part of the compiler are objects. Objects represent anything that can be assigned to a variable in the debris language.

Modules

error
Compilation error handling

function_interface
This module contains a trait ToFunctionInterface which allows to write nice function interfaces by using much less boilerplate to define functions.

hir
High-level intermediate representation

llir
Low-level intermediate representation

memory

mir
Mid-level intermediate representation

objects
Contains all objects which can be used in a debris program.

Structs

CompileContext
The Compilation context stores various information about the current compilation

Config
Holds data about the user specified configuration

DebrisObject
Objects are a central type for the compiler. Basically anything that can be assigned to a variable is an object. This includes numbers, function, modules, and more. It is possible to cast the ObjectPayload to its original value. Because DebrisObject is unsized, it generally only accessed as ObjectRef

Namespace

ObjectRef
This struct is used to pass objects arround

Enums

BuildMode
The build mode for this compilation

OptMode
How to optimize the code

Type
The type of a class object

TypePattern
The enumeration of patterns allowed as function arguments

Traits

ObjectPayload
A trait for values that can be used as debris object payload

ValidPayload

Type Definitions

ObjectProperties
The type of the properties map