Crate debris_core

Core implementation of the debris compiler

This crate contains most of the compilation logic.

Compiling a debris script uses three different intermediate representations: First of all, the text is parsed and a ast-like structure gets produced. This happens in the hir module.

Subsequently, this structure gets lowered into a mid-level intermediate representation. The mir contains debris_objects.

At last the mir will be converted into a low-level intermediate representation. The llir is quite similar to actual minecraft command syntax and contains various nodes for scoreboard operations and more.

A very important part of the compiler are objects. Objects represent anything that can be assigned to a variable in the debris language.

Modules

error
Compilation error handling

hir
High-level intermediate representation

llir
Low-level intermediate representation

mir
Mid-level intermediate representation

objects
Contains all objects which can be used in a debris program.

Structs

CompileContext
The Compilation context stores various information about the current compilation

Config
Holds data about the user specified configuration

DatabaseStruct

DebrisObject
Objects are a central type for the compiler. Basically anything that can be assigned to a variable is an object. This includes numbers, function, modules, and more. It is possible to cast the ObjectPayload to its original value.

InputsStorage
Representative struct for the query group.

ObjectRef
This struct is used to pass objects arround

Enums

BuildMode
The build mode for this compilation

Traits

Inputs

ObjectPayload
A trait for values that can be used as debris object payload

Type Definitions

ObjectProperties