Module debris_core::error

Compilation error handling

Exports the default Result type which is used in debris_core where errors can happen

Structs

LangError
A generic error which gets thrown when compiling

ParseError
Thrown when parsing bad input

SliceOwned
An owned counterpart to the annotate_snippets::Slice struct

SnippetOwned
An owned counterpart to the annotate_snippets::Snippet struct

SourceAnnotationOwned
An owned counterpart to the annotate_snippets::SourceAnnotation struct

Enums

CompileError
A Compile Error

LangErrorKind
Specifies a specific error reason

Traits

AsAnnotationSnippet

Type Definitions

LangResult
A result type which allows quick error throwing since no span and other boileplate is needed

Result
The result type used by most of the core functions