Trait debris_core::function_interface::ToFunctionInterface

pub trait ToFunctionInterface<Params, Return> where
    Return: ValidReturnType,
    Self: 'static, {
    fn to_function_interface(
        &'static self
    ) -> Box<dyn NormalizedFunctionInterface>;

    fn query_parameters(ctx: &CompileContext) -> FunctionParameters;

    fn query_return(ctx: &CompileContext) -> Option<Rc<Class>> { ... }
}

This trait can convert functions into compatible interface functions

Required methods

fn to_function_interface(&'static self) -> Box<dyn NormalizedFunctionInterface>

fn query_parameters(ctx: &CompileContext) -> FunctionParameters

Static method for querying the accepted parameters

Provided methods

fn query_return(ctx: &CompileContext) -> Option<Rc<Class>>

Static method for querying the return type, may be None