Trait debris_core::function_interface::ValidReturnType

pub trait ValidReturnType {
    pub fn to_result(
        self, 
        ctx: &mut FunctionContext<'_, '_>
    ) -> LangResult<ObjectRef>;

    pub fn get_class(ctx: &CompileContext) -> Option<GenericClassRef>;
}

Trait used for converting any valid return value into a LangResult

Required methods

pub fn to_result(
self,
ctx: &mut FunctionContext<'_, '_>
) -> LangResult<ObjectRef>

pub fn get_class(ctx: &CompileContext) -> Option<GenericClassRef>

If possible, returns the type of the return value

Loading content...