Module debris_core::hir

High-level intermediate representation

Parses debris code into a hir. This intermediate representation is very similar to a typical abstract syntax tree, but the some desugaring gets applied.

Modules

hir_nodes

Structs

DebrisParser
The pest parser which can parse the grammar file

Hir
A high level intermediate representation

IdentifierPath
A list of SpannedIdentifiers, can be a dotted path

SpannedIdentifier
Identifies a variable or value based on its span

Enums

Rule

Functions

get_span