Module debris_core::hir

High-level intermediate representation

Parses debris code into a hir. This intermediate representation is very similar to a typical abstract syntax tree, but the some desugaring gets applied.

Modules

hir_nodes
Defines every node used in the hir representation

Structs

DebrisParser
The pest parser which can parse the grammar file

Hir
The hir representation of an input file and all of its dependencies

HirContext
Contains state data that are used during the hir construction

HirFile
This struct stores the high-level intermediate representation of a single file. A HirFile is very similar to a super::hir_nodes::HirModule, but it can store a list of imports

IdentifierPath
A list of SpannedIdentifiers, can be a dotted path

ImportDependencies
Keeps track of all imported modules, uses indexes as keys

SpannedIdentifier
Identifies a variable or value based on its span

Enums

Rule

Traits

HirVisitor
Trait which defines methods that a visitor has to implement