Module debris_core::hir::hir_nodes

Structs

HirBlock

HirFunction
A function, which contains other statements

HirFunctionCall
Any function call, can be dotted

HirInfix
Holds an infix operator combined with its span

HirPrefix
Holds a prefix operator combined with its span

HirPropertyDeclaration
Declaration of a property in a struct definition

HirStruct
A struct definition

HirVariableDeclaration
Holds a variable type declaration like foo: String

HirVariableInitialization

Enums

HirComparisonOperator
Any supported comparison operator

HirConstValue
A constant literal, already parsed

HirExpression
Any expression

HirInfixOperator
Any operator that can be used as an infix

HirItem
Any Item

HirObject

HirPrefixOperator
Any prefix operator

HirStatement
Any statement, the difference to an expression is that a statement does not return anything