Module debris_core::llir::llir_nodes

Contains every node that can be produced in the llir step.

Note that changing any node kind can lead to miscompilations if it isn’t also updated at the optimizers!

Structs

BinaryOperation

Operates on two scoreboard values and stores the result into the target var

Branch

Branches based on a condition

Call

Calls a function

ExecuteRaw

Executes a literal string

FastStore

Stores a ‘fast’ variable

FastStoreFromResult

Stores a ‘fast’ variable from the result of another node

Function

A function node, contains other nodes

WriteMessage

Writes a formatted message

Enums

Condition

Evaluates a condition and returns either true or false

ExecuteRawComponent

A component for a raw execute command. Either string or scoreboard.

Node

Any node

VariableAccess

Denotes how a specific node accesses variables.

VariableAccessMut

See VariableAccess.

WriteTarget

The buffer to write to