Module debris_core::mir

Mid-level intermediate representation

Once the high lever representation has be created, it can be lowered into a mir. The mir is quite simplistic and focuses on control flow.

An input of let a = 3 * 5 + 2 would be represented as a bunch of FunctionCall(...) and look somewhat like this:

Call(
    StaticInt.*,
    parameters: [StaticInt(3), StaticInt(5)],
    return_val: {id: 1, type: StaticInt}
);

Call(
    StaticInt.+,
    parameters: [{id: 1, type: StaticInt}, StaticInt(2)],
    return_val {id: 2, type: StaticInt}
);

The actual value of the returned static ints will be lazily evaluated at llir

A context is used to keep track of all variables that have ben created, as well as their identifiers.

Structs

ItemIdentifier
Identifies a unique template

Mir
A Mid-level intermediate representation

MirBuilder
Visits the hir and creates a mir from it

MirCall
A function call to api functions

MirContext
Keeps track of single context, which can be a function, a loops or something else. As a rule of thumb, everything which has its own namespace is a context.

MirContextInfo
Helper struct which can hold mutable references to the arena and the context

MirGotoContext
Calls a specific context

MirInfo
Struct that is passed around when working with the mir context

MirRawCommand
A raw command

Enums

MirNamespaceEntry
A value in the mir namespace

MirNode
Any node that can be part of the mir representation

MirValue
Any value that is used in the mir compilation and also in the llir

Traits

MirVisitor

Type Definitions

NamespaceArena