Struct debris_core::mir::MirBuilder

pub struct MirBuilder<'a, 'ctx> { /* fields omitted */ }

Visits the hir and creates a mir from it

Implementations

impl<'a, 'ctx> MirBuilder<'a, 'ctx>

pub fn new(
mir: &'a mut Mir<'ctx>,
modules: &[ModuleFactory],
compile_context: &'ctx CompileContext,
code: CodeRef<'ctx>
) -> Self

impl<'code> MirBuilder<'_, 'code>

Implements functionality for working with the attributes mutably

pub fn context_mut(&mut self) -> &mut MirContext<'code>

Returns a mutable reference to the current context

pub fn context(&self) -> &MirContext<'_>

Returns a shared reference to the current context

pub fn context_info(&mut self) -> MirContextInfo<'_, 'code>

Returns a helper struct that contains both a context and the arena

pub fn namespace_mut(&mut self) -> &mut Namespace<MirNamespaceEntry>

Returns a mutable reference to the current namespace

pub fn arena_mut(&mut self) -> &mut NamespaceArena

Returns a mutable reference to the global arena

pub fn arena(&self) -> &NamespaceArena

Returns a shared reference to the global arena