Module debris_core::objects

Contains all objects which can be used in a debris program.

Structs

CallbackFunction
Wrapper, so traits like Eq can be implemented

DynInt
A dynamic Integer

FunctionContext
The context which gets passed to a function

FunctionSignature
A signature describing a single overload of a function

FunctionSignatureMap
A collection of overloads for the same function

ModuleFactory
A wrapper function for functions that return modules

ObjectFunction
A function object

ObjectModule
A module object

ObjectString
A static string object

ObjectType
The type of an object

StaticInt
A static integer object

Type Definitions

TypeRef