Module debris_core::objects

Contains all objects which can be used in a debris program.

Structs

FunctionContext
The context which gets passed to a function

FunctionOverload
A signature describing a single overload of a function

FunctionParameterDefinition

FunctionSignature
A signature containing expected parameters and return type

GenericClass

ModuleFactory
A wrapper function for functions that return modules

ObjBool
A boolean value that is stored on a scoreboard

ObjClass
The class of a value.

ObjFunction
A function object

ObjInt
A dynamic Integer

ObjModule
A module object

ObjNativeFunction
A callable function object for functions declared in debris

ObjNativeFunctionSignature
Created when the mir comes across a function definition, no actual function gets created The Native Function objects get created for each call to such a function signature

ObjNull

ObjStaticBool

ObjStaticInt
A static integer object

ObjString
A static string object

Enums

FunctionParameters
Decides which arguments a function can accepts

Traits

HasClass
Marks objects that have a class

Functions

match_parameters

Type Definitions

ClassRef
A reference to a class

FunctionSignatureRef

GenericClassRef