Trait debris_core::ObjectPayload

pub trait ObjectPayload: ValidPayload {
    pub fn generic_class(&self, ctx: &CompileContext) -> GenericClassRef { ... }

    pub fn get_property(&self, _: &Ident) -> Option<ObjectRef> { ... }

    pub fn as_function(&self) -> Option<&ObjFunction> { ... }
}

A trait for values that can be used as debris object payload

The private AsAny trait is auto-implemented

Provided methods

pub fn generic_class(&self, ctx: &CompileContext) -> GenericClassRef

The class specific to this object. Can contains some extra generics

By default defers to the HasClass::class implementation

pub fn get_property(&self, _: &Ident) -> Option<ObjectRef>

May be overwritten by distinct payloads which carry properties

pub fn as_function(&self) -> Option<&ObjFunction>

Interface any object can implement to be treated as a function

Loading content...