Trait debris_core::ObjectPayload

pub trait ObjectPayload: ValidPayload {
    fn memory_layout(&self, ctx: &CompileContext) -> MemoryLayout;

    fn generic_class(&self, ctx: &CompileContext) -> GenericClassRef { ... }

    fn get_property(&self, _: &Ident) -> Option<ObjectRef> { ... }

    fn as_function(&self) -> Option<&ObjFunction> { ... }
}

A trait for values that can be used as debris object payload

The private AsAny trait is auto-implemented

Required methods

fn memory_layout(&self, ctx: &CompileContext) -> MemoryLayout

Returns the memory layout of this specific object This method is usually only called once

Loading content...
Provided methods

fn generic_class(&self, ctx: &CompileContext) -> GenericClassRef

The class specific to this object. Contains additionally to the class generics and the memory layout

Per default contains no generics

fn get_property(&self, _: &Ident) -> Option<ObjectRef>

May be overwritten by distinct payloads which carry properties

fn as_function(&self) -> Option<&ObjFunction>

Interface any object can implement to be treated as a function

Loading content...